adjusted modularity of components
